Rename EmailProviderViewController to SetupImapViewController
